YAY-11 refactor: move SliderRoot to folder and extract history slider into standalone component
