+++ src/shared/components/Slider/SliderRoot/SliderRoot.tsx
