+++ src/shared/components/Slider/index.ts
@@ -1,6 +1,7 @@
-export * from './SliderRoot'
+export * from './SliderRoot/SliderRoot'
 export * from './SliderArrow/SliderArrow'
 export * from './SliderContent/SliderContent'
 export * from './SliderSlide/SliderSlide'
 export * from './SliderDots/SliderDots'
 export * from './useSlider'
+export * from './Slider'
